# apps/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Local Market API",
    "description": "API publique de découverte des exploitations et de leurs offres.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "operationId": "login",
        "tags": ["auth"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "format": "email",
                    "maxLength": 254,
                    "type": "string"
                  },
                  "password": {
                    "minLength": 8,
                    "maxLength": 256,
                    "type": "string"
                  }
                },
                "required": ["email", "password"]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "email": {
                      "format": "email",
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "format": "date-time",
                      "type": "string"
                    }
                  },
                  "required": ["userId", "email", "displayName", "expiresAt"]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "format": "uri",
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {}
                    }
                  },
                  "required": ["type", "title", "status", "code", "detail", "requestId", "errors"]
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "format": "uri",
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {}
                    }
                  },
                  "required": ["type", "title", "status", "code", "detail", "requestId", "errors"]
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/session": {
      "get": {
        "operationId": "getSession",
        "tags": ["auth"],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "email": {
                      "format": "email",
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "format": "date-time",
                      "type": "string"
                    }
                  },
                  "required": ["userId", "email", "displayName", "expiresAt"]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "format": "uri",
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {}
                    }
                  },
                  "required": ["type", "title", "status", "code", "detail", "requestId", "errors"]
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/carts/{cartId}": {
      "get": {
        "operationId": "getCart",
        "tags": ["carts"],
        "parameters": [
          {
            "schema": {
              "minLength": 40,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "cartId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/carts/{cartId}/items": {
      "post": {
        "operationId": "addCartItem",
        "tags": ["carts"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "listingId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "quantity": {
                    "pattern": "^\\d+(?:\\.\\d{1,3})?$",
                    "type": "string"
                  }
                },
                "required": ["listingId", "quantity"]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "minLength": 40,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "cartId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/carts/{cartId}/items/{itemId}": {
      "patch": {
        "operationId": "updateCartItem",
        "tags": ["carts"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "pattern": "^\\d+(?:\\.\\d{1,3})?$",
                    "type": "string"
                  }
                },
                "required": ["quantity"]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "cartId",
            "required": true
          },
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "itemId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "operationId": "deleteCartItem",
        "tags": ["carts"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "cartId",
            "required": true
          },
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "itemId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/checkout": {
      "post": {
        "operationId": "checkout",
        "tags": ["checkout"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cartId": {
                    "type": "string"
                  },
                  "email": {
                    "format": "email",
                    "type": "string"
                  }
                },
                "required": ["cartId", "email"]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "minLength": 8,
              "maxLength": 128,
              "type": "string"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/guest-orders/{token}": {
      "get": {
        "operationId": "getGuestOrder",
        "tags": ["orders"],
        "parameters": [
          {
            "schema": {
              "minLength": 32,
              "maxLength": 100,
              "type": "string"
            },
            "in": "path",
            "name": "token",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/order-claims/request": {
      "post": {
        "operationId": "requestOrderClaim",
        "tags": ["orders"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "format": "email",
                    "type": "string"
                  },
                  "reference": {
                    "type": "string"
                  }
                },
                "required": ["email", "reference"]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/order-claims/confirm": {
      "post": {
        "operationId": "confirmOrderClaim",
        "tags": ["orders"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "minLength": 32,
                    "type": "string"
                  }
                },
                "required": ["token"]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
...
}